simplify tests/cli/vyper_json
