+++ tests/cli/vyper_json/test_get_inputs.py
