+++ tests/cli/vyper_json/test_interfaces.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 from pathlib import PurePath
 
 import pytest
